backend/requirements.txt
fastapi
uvicorn
websockets
faster-whisper
requests
python-dotenv
python-multipart
ultralytics
opencv-python-headless
scipy
